Move gios coordinator to separate module
