+++ tests/components/gios/__init__.py
@@ -37,18 +37,19 @@ async def init_integration(
 
     with (
         patch(
-            "homeassistant.components.gios.Gios._get_stations", return_value=STATIONS
+            "homeassistant.components.gios.coordinator.Gios._get_stations",
+            return_value=STATIONS,
         ),
         patch(
-            "homeassistant.components.gios.Gios._get_station",
+            "homeassistant.components.gios.coordinator.Gios._get_station",
             return_value=station,
         ),
         patch(
-            "homeassistant.components.gios.Gios._get_all_sensors",
+            "homeassistant.components.gios.coordinator.Gios._get_all_sensors",
             return_value=sensors,
         ),
         patch(
-            "homeassistant.components.gios.Gios._get_indexes",
+            "homeassistant.components.gios.coordinator.Gios._get_indexes",
             return_value=indexes,
         ),
     ):
